WIP fix(git): set cwd to root git dir
There are still some situations where some VGit features will fail when
the cwd doesn't match the directory used when opening the file. For
example,

This commit reverts a62ad74 ("fix(git): cwd is sometimes required") and
provides a more robust solution.

To make this work in all situations, git -C is always set to the root
git directory for a file, and the filename is always the full relative
path of the file from the root git dir. This approach is agnostic to the
cwd of vim, so should work anywhere, no matter what directory the file
was opened from or the current directory.

The "git -C <path>" prefix can execute git commands as if they were in
<path> instead of the current working directory. This is important for
VGit where :pwd may not be the git root dir.

Currently GitObject calls the Git(...) constructor with a
fs.dirname(filename) arg to set self.cwd (which then sets the <path> for
git -C). This works in most cases as long as we were in the correct
directory when we opened the file.

Instead, set self.cwd and git -C <path> to be the output of
"git rev-parse --show-toplevel" which is the absolute path to the root
git directory of the current file in all cases. The filename is always
the relative path from the root git dir.

TODO:
- need to update + check other screens - need to be amended to accept
  full path there
- update filename just to be full relative path everywhere - then remove
  filename
    - or just use the full path and then do fs.make_relative with
      self.cwd (aka root git dir) when doing git cmds
- test in multi-level nested/embedded git dirs
    - cd into another git dir and check
- if not inside git dir, we use fallback_cwd from git config - what is
  this, do we need to keep it?
    - what happens outside of git dir + and using a fallback with these
      changes?